modifying the if else statement.
